- Scenes pass their settings file into their action, so custom settings can be used per scene - Added precompiler > Update NuGet packages > Update nd-codegen > Run nd-codegen to update entities > Switch long transitions to floats > Switch between local and remote IP based on build vs publish

+++ NetDaemonWrapper/Lighting/LightExtensions.cs
@@ -17,7 +17,7 @@ namespace NetDaemonWrapper.Lighting
             _light.TurnOn(rgbColor: color, brightness: _colorBright.brightness, transition: 1);
         }
 
-        public static void Set(this LightEntity _light, ColorBright _colorBright, int _transition)
+        public static void Set(this LightEntity _light, ColorBright _colorBright, float _transition)
         {
             if (_colorBright.brightness == 0)
             {
